rename findapal package to pals
