+++ app/src/main/res/layout/fragment_incoming_pal_requests.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".views.findapal.IncomingPalRequestsFragment">
+    tools:context=".views.pals.IncomingPalRequestsFragment">
 
     <TextView
         android:id="@+id/palRequestTitle"
