+++ app/src/main/res/navigation/nav_graph.xml
@@ -79,12 +79,12 @@
     </fragment>
     <fragment
         android:id="@+id/findAPalFragment"
-        android:name="com.example.packpals.views.findapal.FindAPalFragment"
+        android:name="com.example.packpals.views.pals.FindAPalFragment"
         android:label="@string/FindAPal"
         tools:layout="@layout/fragment_find_a_pal" />
     <fragment
         android:id="@+id/incomingPalRequestsFragment"
-        android:name="com.example.packpals.views.findapal.IncomingPalRequestsFragment"
+        android:name="com.example.packpals.views.pals.IncomingPalRequestsFragment"
         android:label="@string/IncomingPalRequests"
         tools:layout="@layout/fragment_incoming_pal_requests" />
 </navigation>
